change record StatReview zscore from Float to BigDecimal

+++ src/main/java/com/zscore_api/zscore_api/record/StatReview.java
@@ -1,4 +1,6 @@
 package com.zscore_api.zscore_api.record;
 
-public record StatReview (Integer gameId, Integer pubId, String gameTitle, String pubName, Integer score, Float zscore) {
+import java.math.BigDecimal;
+
+public record StatReview (Integer gameId, Integer pubId, String gameTitle, String pubName, Integer score, BigDecimal zscore) {
 }
